Optimize release task
Prior to this commit we were reuploading all commits in the last 24 hours
This was due to a misunderstanding of how Puppet pipelines worked.This commit
uses the current repo ( which is reset by pipelines to the commit that
generated the webhook ). This means only the last commit is checked for changes
as each commit will cause a payload to kick off the pipeline jobs. This should
speed up deployments as well as cause less traffic and API calls in learndot.