Check if file is empty in addtion to non-existant
Prior to this commit we only checked if the file existed. This commit
checks if the file is "blank" and simply a place holder. This should prevent
empty lines being uploaded to learndot and the formatting to be weird. These
fields are not required in the interface and thus shouldn't be required here
when place holders are present.